:recycle:  deduplicate code printer-legacy section metadata
