+++ packages/printer-legacy/src/graphql/object.js
@@ -1,16 +1,11 @@
 const {
   object: { hasMethod },
-  graphql: { getTypeName, getFields, isDeprecated },
+  graphql: { getTypeName, getFields },
 } = require("@graphql-markdown/utils");
 
-const {
-  printSection,
-  HIDE_DEPRECATED,
-  SHOW_DEPRECATED,
-} = require("../section");
+const { printSection, printMetadataSection } = require("../section");
 const { printCodeField } = require("../code");
 const { MARKDOWN_EOL } = require("../const/strings");
-const { OPTION_DEPRECATED } = require("../const/options");
 
 const printInterfaceMetadata = (type, options) => {
   if (hasMethod(type, "getInterfaces") === false) {
@@ -21,47 +16,15 @@ const printInterfaceMetadata = (type, options) => {
 };
 
 const printObjectMetadata = (type, options) => {
-  const values = getFields(type);
-
   const interfaceMeta = printInterfaceMetadata(type, options);
+  const metadata = printMetadataSection(
+    type,
+    getFields(type),
+    "Fields",
+    options,
+  );
 
-  switch (options.printDeprecated) {
-    case OPTION_DEPRECATED.GROUP: {
-      const { fields, deprecated } = values.reduce(
-        (res, arg) => {
-          isDeprecated(arg) ? res.deprecated.push(arg) : res.fields.push(arg);
-          return res;
-        },
-        { fields: [], deprecated: [] },
-      );
-
-      const meta = printSection(fields, "Fields", {
-        ...options,
-        parentType: type.name,
-      });
-      const deprecatedMeta = printSection(deprecated, "", {
-        ...options,
-        parentType: type.name,
-        level: "",
-        collapsible: {
-          dataOpen: HIDE_DEPRECATED,
-          dataClose: SHOW_DEPRECATED,
-        },
-      });
-
-      return `${meta}${deprecatedMeta}${interfaceMeta}`;
-    }
-
-    case OPTION_DEPRECATED.DEFAULT:
-    default: {
-      const metadata = printSection(values, "Fields", {
-        ...options,
-        parentType: type.name,
-      });
-
-      return `${metadata}${interfaceMeta}`;
-    }
-  }
+  return `${metadata}${interfaceMeta}`;
 };
 
 const printCodeType = (type, entity) => {
